[IMP] Separando classes em arquivos diferentes
